+++ src/org/semanticweb/HermiT/datatypes/floatnum/NoNaNFloatSubset.java
@@ -45,7 +45,7 @@ class NoNaNFloatSubset implements ValueSpaceSubset {
     @Override
     public boolean containsDataValue(Object dataValue) {
         if (dataValue instanceof Float) {
-            float number=(Float)dataValue;
+            float number=((Float)dataValue).floatValue();
             for (int index=m_intervals.size()-1;index>=0;--index)
                 if (m_intervals.get(index).contains(number))
                     return true;
